Moves logic to `NonFinalizedBlocksListener::spawn`,moves channel buffer size to a constnat, and adds correctness notes.
